requirements.txt
load_dotenv 
openai 
streamlit
korean_lunar_calendar
pytz
pandas
sqlalchemy
pymysql
mys
ql-connector-python